+++ source/app/util/custom-protocols.ts
@@ -13,7 +13,7 @@
 
 import type LogProvider from '@providers/log'
 import { protocol } from 'electron'
-import { promises as fs } from 'fs'
+import { promises as fs, constants as FSConstants } from 'fs'
 import path from 'path'
 
 // Reference if we need more: https://www.iana.org/assignments/media-types/media-types.xhtml
@@ -66,6 +66,25 @@ export default function registerCustomProtocols (logger: LogProvider): void {
       if (/^\/[A-Z]:/i.test(pathName)) {
         pathName = pathName.slice(1)
       }
+
+      if (pathName.startsWith('//')) {
+        // makeValidUri ensures that a network share path in the renderer
+        // process results in four slashes here, meaning pathName can start with
+        // two slashes. On Windows, apparently it works to simply transform
+        // safe-file:////remoteHost/file.png -> /remoteHost/file.png. On macOS,
+        // provided the share is currently mounted, we have to transform it to
+        // /Volumes/remoteHost/file.png, and on Linux we won't do s***.
+        if (process.platform === 'win32') {
+          pathName = pathName.slice(1)
+        } else if (process.platform === 'darwin') {
+          pathName = `/Volumes/${pathName.slice(2)}`
+        } else {
+          return new Response(null, { status: 404, statusText: 'File not found' })
+        }
+      }
+
+      // File must be visible to the process and readable
+      await fs.access(pathName, FSConstants.F_OK | FSConstants.R_OK)
 
       const fileBuffer = await fs.readFile(pathName)
       return new Response(fileBuffer, {